* Partially fixed FSReference>>renameAs:. It now works for disk filesystems. * added FSReference>>moveTo:
WARNING: this introduces a slight change into the API. #renameAs: now is a shortcut to simply change a file's name while staying in the same parent directory. As such, it now takes a string (used to take a path or string). What renameAs: used to do (move to an arbitrary other location) is now done by #moveTo: which takes a target reference.