avoid complete history traversal by VersionHistoryWalker
Use an algorithm that marks versions similar to the merge-base algorithm.
Stop early if no version in the queue was not yet reached from a sink.
Prioritizing versions by timeStamp is now mandatory for this to work correctly.
For Git commits the committer timeStamp must be used because
author timeStamps need not be chronological (--> cherry-pick, rebase).

Remove accessors for sourceVersions and sinkVersions because they were unused.
The existing high-level methods should be used instead.