fix: created bogus merge commits with two identical parents
When you 'merge into memory' the latest version of the active branch
and the 'last loaded' version was on another branch, new head of the
active branch is added as an additional parent.
Not all code paths removed that additional parent if the active branch
was still at the same version at the time of committing, so a commit
could be created with two identical parents (same hash).