copy trees before modifying them
This fixes an issue with GitFilesystems that influenced another
because of the objectCache in GitRepository:
A tree is read from the repository and cached.
The tree is modified by a filesystem, which previously
changed the cached tree object.
A new filesystem is created on the old tree, but the updated
entries would show up because the cached tree was modified.