add compare with working copy and hidden branch activator
From the shifted branch list yellow button menu you can now
make a branch active without affecting the objects in a working copy.
Usually, one should not do this, but it might be necessary to properly
hook up already loaded packages to a newly added Git repository.
To reproduce this in git, one could do this:
  git checkout --detach
  git reset --soft <commit-ish>
  git checkout <commit-ish>
...but usually nobody copies a .git directory into a plain working tree
and then determines that this working tree is based on another commit.