Prevent surprises if a user changes selections during prompts
While some information must be entered, the user might
change the project selection or branch selection. This should not alter
the target of the ongoing operation that is currently prompting.
For example, if you chose to add a package to a project and then
changed the project selection while the categories prompt is displayed,
the package would previously have been added to the newly selected
project instead of the originally selected one.