Fix: could not properly resolve a conflict in SquotChangedArtifactsChooser
The dialog would refuse to accept, always complaining that a conflict were not resolved yet.
The artifactDiffs are only modified once the #selectedPatch is retrieved.
This patch in turn will only be non-empty if the dialog has been accepted.
The reason not to change diffs/resolve conflicts immediately is to allow canceling.