Fix: load order changes not preserved in all cases
Rebuilding a partial original and a patched collection will not work
because the (unchanged) elements not available in the diff snippets must
be replaced by placeholders and those will open up new inadvertent
possibilities for matches by the diff algorithm.
You have to stick to the available snippets and create new chunks and
snippets out of them.