Refactor diff manipulation through SquotDiffNode
Instead of manipulating the diffs via adjustedBy: when the changes chooser is accepted
directly manipulate the diff when nodes are activated and deactivated.
To define the manipulation, blocks are provided when the nodes are created.
The nodes are created from the diffs that will be manipulated,
so the diffs provide the blocks for their own manipulation.