Fix flaky chronological order test
The chronological order uses the committer timestamp.
Before DataAndTime got updated in Squeak trunk to be 
UTC-based, it seems to have had nanosecond precision, 
which it now has not because the new DateAndTime 
stores microseconds since an epoch. It happened more often 
than before that two commits got the same "now" 
timestamp when they were created during the test.
Use the supplied DateAndTime also for the committer 
timestamp to circumvent the issue.