Do not require metadata files when the file store is patched
Just write the right (patched) version of the patch, it has metadata ready.
But we have to make sure that the right side of an artifact diff is
up to date, since the patch could have been modified in the save dialog.

Moved SquotArtifact>writeWith: to the trait TSquotArtifact because
the right side of an artifact diff can also be a SquotLiveArtifact.