Fix tests for conflict resoltion with the tools
Because the right side of the diffs is updated now,
we can no longer assert that it is still the same as before.
This assertion did not make much sense anyway
since the subject of these tests are the tools and
the interaction with the diff nodes,
and not how patches work.