Pull addGlobalObjectsToStore: out of patch loading
Invoking the method on a patch requires artifact diffs to implement
the protocol of artifacts. This has worked so far, but it is a smell.

Due to the changed order in some operations, the orphan merge scenario
had to be adapted: the package of the version to be merged must be
purged before the merge, or the residual package will be picked up
in preparation of the actual merge.