WIP: Merges are not diffs, nor are conflicts, so separate them
SquotPatch should not respond to hasConflicts because a patch
should only apply changes, so the conflicts must already be resolved.
Resolving conflicts is done during a merge.

SquotMerge is the entry point to the new code.

Since merging might need a separate tool soon, separate the test cases
for merging from SquotChangedArtifactsChooserTest into SquotMergeToolTest.

Exposes error in object graph manipulation: new objects behind an added one
do not get added to the nodes of an object graph.

Work in progress: conflicts at the level of whether containers include an artifact
or not are separated, but not conflicts in object graphs. These are not detected
right now, which makes additional tests fail.