Replace SquotArtifactDiffWithContentAndBase by SquotArtifactMerge
SquotObjectGraphMerge now contains code previously found in
SquotDiffBuilder for merging.
Merging of objects is still not implemented (except for packages).
The new code only goes as far as the old code did.

Change SquotMonticelloMergeDiff to not replace conflicts by modifications.
Instead, the MCConflicts are resolved.

Introduce an extra tool subclass to handle merges instead of patches.

Refactor SquotChangedArtifactsChooserTest such that the test cases
can still be inherited, but methods that produce input or access the 
output of the operation can be overridden. This allows to reuse the 
tests for merges as well as patches.

Extract more general tool testing utilities into superclasses of SquotDiffExplorerTest.