Use proper merge objects instead of patches
Each merge pertains to an object. Most of such objects are somehow
containers or collections. These can have conflicts about whether an
element with a specific key is part of the collection or not (like a tree
conflict about a file name in file-based VCS). Moreover there can be
agreeing modifications to the collection, such as adding or removing
elements. All elements need to be merged themselves.

Removed hasConflicts from patches and most diffs.

Repurpose selector "squotDiffAgainst:" for "flat" comparisons that 
don't go into the object graph. There is still code duplication with 
the deep comparisons involving the SquotDiffBuilder.

Split SquotObjectDiffTest into SquotObjectGraphDiffTest.

Do not support changing of start objects in graphs for the moment.
Patching tests are still broken, and probably so is materialization.