Acknowledge bug in shadow patching, relax assertion accordingly
The patched graph is inconsistent, if not plain wrong,
about the modified objects. Several editions of the same
object can be referenced in the intensional graph of shadow
objects, while the (extensional) graph itself does not even
know the patched editions.

Note that this is irrelevant for the patching of live objects
because these are modified in-place. There exist no copies
of them for other editions, there is only the one loaded edition.
This trivially solves inconsistencies. Also the patched shadows
get materialized right away. They are not correctly stored in
the patched graph of shadow objects at the moment.