Compute updated graph of resolved merge
To that end, partially fix shadow graph patching. 
Implements such patching without the use of the
SquotObjectPatcher helper object. The references in the 
graph are still not updated!

Change keys in SquotObjectGraphMerge to object names
instead of the objects themselves. This makes it consistent
with the keyOf: method.

Note: one comment states "value objects are not merged",
but packages are currently treated like value objects, except
that they can have diffs and are indeed merged...

startDiff assignment in SquotShadowGraph reverted because
the shadow graph is supposed to be immutable.