Fix: nil "inherited" names of garbage-collected objects sometimes
The objectNames dictionary of the registry's graph has weak keys,
which become nil when they are collected. If the collected object
had its WeakKeyAssociation at a suitable index in the objectNames
hash table, the stale association was picked up when the registry
was asked for the name of nil. The WeakKeyDictionary does not catch
this case.

The new canHaveNameAssignedBySquot is different from 
squotShouldKeepIdentity. Even objects that don't need
names can get one as the start object of a graph. This is
the case for PackageInfo/SquotPackageShadow, for example.
But trivial objects must not be tracked as start objects.
It does not make sense, since they cannot be changed anyway.