Fix: 'Browse full' in change chooser signalled errors
Implementation was only valid for MCPatchOperations.
This delegates it to the contents of the node, which can
be diffs or conflicts of any kind.

Implement browse action for all MCDefinitions.

There is a message semantics ambiguity though:
#browseFromSquot: is also used for browsing snapshots
from SquitBrowser. 'Browse full' should rather browse the
live objects as expected from the rest of the Squeak IDE.