Move path validation responsibility to stores
It was formerly with the artifacts, but it did not make much sense.
The stores put constraints on the paths, so they should be responsible.
Repositories must also be able to validate paths as they are expected
to use some kind of store to access the VCS backend.

When the serializer of a package is changed from the Git browser,
it will now modify the working copy wrapper of the package artifact,
invoking the path validation and fixup for image store and repository explicitly.

Fix tests where I have been lazy and used shortened paths 
that were formerly extended by the serializer.