Fix: metadata was not used during capturing in image store
This broke the workspace tracking test because now the owner
of the tracked system window was not excluded. The store thus
attempted to capture the whole world leading to all kinds of errors.
All kinds because the errors depended on the tools and other morphs
you have open in the world...