Move environment handling to package classes
Artifacts with environments (when loaded) are a specialty.
Currently it only applies to packages. So the image store should
not need to deal with this specialty. The image store should
instead always activate the metadata for the artifact, so
shadows, materializers etc. can access such special information.

Remove double-initialization of the environment by addLoadedPackageInfoNamed:at:serializer: