Make the current environment the default in PackageInfo
The default environment returned by PackageInfo is changed
to the current environment (instead of the global default environment)
to make it possible to influence the target environment from 
the invoking context. This is useful if the working copy is not yet
initialized and does not store an environment for the package yet.