New cherry pick implementation with working copy operations
Extract operations such as merging and cherry picking out of
SquotWorkingCopy. Saving shall also be extracted but it is not
done yet (the cherry pick operation still uses the old save methods
on SquotWorkingCopy instead of the operation, which is unfinished).

The goal is to make such operations more easily extendable. Adding
more and more arguments to existing methods is not an option.
For example, cherry picking should preserve the original author
(ref #194), but this needs another input variable for the save operation
in addition to the patch and the message.

Abandon block callbacks and use asynchronous messaging with Promise
instead. This naturally solves the problem that modal dialogs are defunct
when their UI process is terminated (which unwinds the stack with the
ongoing operation).

Since the save operation is not ready yet, the SquotChangedArtifactsChooser
has been retrofit with a Promise interface. It provides a hack
to recapture the modal invocation in the new UI thread if the
stack is unwound. There are no test cases for this hack.
The view of the chooser is made an instance variable, because
it does not make sense to have multiple views for the tool,
and it is needed to run the modal loop all but the first time.

SquitBrowser functions as service requestor for the cherry pick operation.
It will request on behalf of the operation whether the pick should be
reverted if the save is cancelled.

SquotToggles provides feature flags. In case the new cherry pick
implementation renders cherry picking unusable somehow, the old
behavior can be restored through it.