Do not patronize the store of the working copy when applying patches
For example, a patch that added the SquotUnrecordedFilesArtifact would
leave behind its path in the load order if the patch was applied to the
working copy and not to the store.

Ideally there were some load order sanitization in the setter of the store,
but delegating to the store entirely also gets the job done.

Another issue was that previously the object registry of the store was
not activated if the diffs were applied via the working copy.