Fix: unsaved changes were silently saved during cherry picks
The diffs had different bases (the working copy) than
what they were applied to (the base version), leading to incorrect
patched versions (with the working copy changes included).
The problem only occurred with the file-based repositories
because they use the #right artifact of an artifact diff.

"Rebase" the diffs when the save operation is prepared.