Add feature toggle for Git repo unit-of-work improvement
The plan is to create a unit-of-work object for each larger
Git operation, such as fetching, pushing, writing new versions
etc. and keep open the file handles of pack files during these
operations. Also the configuration files would only be read once.

This is supposed to alleviate some of the slowness of reading many
Git objects from the repository.