Load commit parents within a single unit of work
During a walk, the parents will have the unit of work as
the repository, reusing the a single one for the whole walk.

Had to relax the merge base tests such that they find
"equal" commits instead of the same GitCommit object.
As far as Git is concerned, these are still the same commits, of course.