Remember pack files in unit of work and keep them open
This speeds up reading of objects quite noticably.
This optimization was removed in GitRepository because when the
set of pack files changes, the new pack files must be picked up.
But during a unit of work, we can assume that no new pack files will appear,
except if the unit of work writes them itself.