Fix repo and historian comparison; rename private repo accessors
Comparison was broken because a GitRepository was compared to
a GitUnitOfWork.

It would be hard to come up with a reasonable name for accessing
the GitRepository (not a unit of work), other than gitRepository.
Unfortunately, this private accessor is not as private as it should be.
The reason is mainly that the SquitBrowser was built quickly, skipping
some architecture concerns.