Share object cache between units of work
After all, the objects in a repository are immutable.
And the Git implementation is reasonably free of bugs by now,
so the Smalltalk Git objects are not corrupted all the time
and can be kept.

To do: evict objects from the cache eventually,
particularly do not hold onto large blobs in the image...