Git objects refer to repository again, but access unit of work through it
This should make sure that a unit of work can be retired
even though the objects that were loaded during it are still around.
They can be used again in a different unit of work, if one was started.

Downside: need to access the DynamicVariable for the current 
unit of work for every operation in it.