Distinguish canHaveNameAssignedBySquot and shouldHaveNameAssignedBySquot
Actually the implementation is the same for basic objects. Maybe the
methods can be conflated in the future.
Currently, canHaveNameAssignedBySquot is sent to shadows and to
primitive and bit objects held in SquotPrimitiveValue, to check whether
it is appropriate to assign an identity.
Whereas shouldHaveNameAssignedBySquot is supposed to be sent
to live objects to find out whether they are value objects (or rather not).