Slightly improve graph encapsulation during capturing
Let the graph deal with the names of objects and their shadows.
Reduces code duplication and routes additional name assignments
through the assertion in addSHadow:of:.

Split basic testPackageInfo because it would remember the
dummy name for the PackageInfo when it is the start object.
Subsequently the assertion in addShadow:of: would fail when
the PackageInfo is captured a second time, but not as the start object.