Prevent looking up the unit of work again to resolve a symbolic ref
The unit of work is in charge of resolving something, so it should not
be necessary to also activate it in the dynamic environment.

Centralize the responsibility of following any symbolic refs in the
unit of work and relieve the ref stores from doing it. Ref stores
should only do a flat look up of any ref and the signature can be
either a hex hash (looked-up ref is a proper ref) or another ref
(looked-up ref is a symbolic ref).