Fix repeated fetch question when the fetch fails
When an error happens while the dialog is open,
the invocation will be curtailed and the dialog cancels.
This triggered the method again that just failed, and involved
showing the fetch question again in this case.

Guard against double execution of the continuation by using a Promise
instead of an event. Promise blocks are by definition evaluated only once.

Ref #181