Do not add remote if fetch failed and cancelled
Copied/extracted error handling from SquitBrowser into an
interactive operation for the fetch.

Running the operation returns a promise. If the operation is
cancelled or if the fetch is curtailed (i.e. cancelled via the debugger)
then reject the promise.

Rejecting the promise will remove the newly added remote in this case.

Ref #181