WIP: Refactor push workflow w.r.t. remote error handling
Several other functions currently broken due to control flow mismatches.
(Promises vs. modal dialogs vs. separate process).

Turns out that debugging the control flow is hard with Promises.
It also is in JavaScript, so this should not be a surprise.
Since this all got too complicated, switch the approach once more:
Write the workflow as simple and straightforward as possible.
Use exceptions for deviations from the standard workflow.
Try to treat asynchronous GUI requests in the operation as if they were
synchronous. Run the operation in another process so we can wait for input
without blocking the UI process.

Refactor some SquitBrowser methods to operate on historians
instead of branchname + repository.

Resolve actionRemoteAdd promise with the SquitAddRemote object
to get information about the new remote in the sender method.

Handle exception when attempting to push a non-real ref.