Fix possible inconsistency in displayed commit list
If two repositories contained the same commits and you
switch between their projects, the optimization to not
reload the commit list if the same commit is still on
top caused that the commits that referred to the old
repo were displayed. Subsequently, there were unit of
work failures because a unit was opened for the wrong
repository.