Fix: missing changes in merge commits
For example, another branch has a package that is not on
your current branch, but you have it already loaded somehow,
but not tracked in the working copy.  During the merge,
nothing changes in the image because the package is already
loaded. The package is just added to the working copy.  But
when saving the merge, the package must appear as added.
Before this commit it did not. The reason is that I wanted
to be smart and exclude uncommitted changes that have
nothing to do with the merge at hand.  But for now I'll just
try to be smart later and go with something that is not
broken.