Fix: author timestamp changed during cherry picks
This error was revealed by making GitStamps immutable. The error was
previously masked by a mutated original GitStamp during the tests.

SquotSaveOperation is used both standalone and during SquotCherryPickOperation.
Standalone, the timestamp must be the current time when the version is stored.
During cherry picks, the timestamp must not be the current time.

Ensure that the file example repository has a default value for version timestamps.

Ref #281