Fix: merged load order could contain removed artifact's path
The auto-resolution of load order conflicts takes all paths from both
sides, but that contradicts a merge resolution where an artifact is
actually removed (or stays removed) during the merge.

There are more problems to just merging the collection of strings:

Diff3 new
	file0: #(a b c d e);
	file1: #(b c d e a);
	file2: #(b c a d e);
	diffClass: HuntMcilroyDiff;
	merge: true
"=> {#ok->#(#b #c #a #d #e #a)}"
	
Note that #a is now doubled in the example. This is not yet
addressed by this commit. There is a missing invariant: an artifact
never appears twice at the same path.