Start implementation of unloaded artifacts, ref #130
Include unloaded artifacts in the snapshot of working copies.
In general, compare with and patch the working copy, not its store.

SquotWorkingCopy>>applyDiff: is implemented via TSquotMutableObjectContainer>>#applyDiff:

Activate the object registry of SquotImageStore around additional
method calls because some higher-level sends go to the working copy now.

SquitPackageChooser: Do no longer assume that the store
will contain all the packages of the project.

In my Squot working copy, the files that do not belong to the packages
(SquotUnrecordedFilesArtifact) are now an unloaded artifact.