Code cleanup (continued): differentiate hasChanges, mayAffectArtifact
Allow arbitrary diffs to be added to a patch, even diffs without changes.
Pull hasChanges check out so the business logic can decide about that.
For saves, only diffs that change the snapshot are relevant.
For loads and merges, also diffs that change the working copy state are relevant.

SquotArtifactLoad is created in disabled state already when an unloaded
working copy artifact is compared to a regular artifact and there are no changes.

Ignore objectsReplacedByNames property of metadata since older repositories
and repositories without any Squot metadata will not have this property.
Do not let unchanged artifacts appear during checkouts because of this.