WIP: Fix toggle load GUI in changes chooser
Display load status and action in root node title and at the top of the text.
That means the titles are not static anymore.
Moreover, the titles can be Texts now (which requires String conversion sometimes).
Use temporary variable highlighting for the color of deactivated diff nodes,
so the deactivated nodes can hopefully be read in themes with gray backgrounds.

Allow spontaneous loading of artifacts being modified.

Move isArtifactNode check for toggle load menu extension,
so it is more obvious when these items appear and when not.

Move conflict highlighting (bold) to SquotDiffNodes
to have fewer places that apply formatting.
(There is SquotDiffNode, the tool, SquotDiffNodeTreeWrapper, the node contents, ...)

Unfinished:
- Merges always assume that the working copy side is loaded.
- Loading during merges is still unsupported.
- All artifacts appear as "not loaded" when saving new versions.

Ref #130