Fix: all artifacts appeared as unloaded when saving
Apply the load status in the save operation, since it is bound to the
working copy. All paths that lead to the save dialog should pass this
at some point.

In SquotSaveOperation, factor out the side effect of initializing the
patch during the validation of the input variables. This should be
easier to understand.

This makes it necessary to split the validation procedures for full
saves and projected cherry pick saves. The patch for a cherry pick
save is only determined after the merge has happened, so not at the
start of the cherry pick operation.

During saves, the right side of the patch represents the working copy.
Hence rename leftIsLoaded to isLoadedInWorkingCopy.

Under some circumstances it happened that the in-memory version
snapshots for tests contained SquotWorkingCopyArtifacts. Prevent this
by putting only the unwrapped artifacts in the (lazy) snapshot.

Ref #130