Allow loading of unloaded artifacts during merges
Merge into the working copy directly, not into a snapshot.

Implement merging dispatch for SquotArtifactLoad.
Removals override loads and additions are mutually exclusive with them.

Proliferate distinction between changes [to snapshots] and
affecting an artifact... for the sake of keeping loads in merges.

More and more often it is easier to test too much than to mock
something. This is because the objects interact so much and the
mocks would have to reimplement half of the real behavior.

Ref #130