Revise: working copy lazySnapshot should contain load status of artifacts
...so merges and cherry picks can see whether artifacts are loaded or not.

This reverts a change from 1fd4b74.
The purpose of the change was to prevent in-memory versions
from having SquotWorkingCopyArtifacts in the snapshot.
Instead, ensure this by unwrapping working copy artifacts when
they are captured. Change saving such that it immediately
captures all artifacts in the working copy.

Refactor cherry-picking accordingly to create the save operation later.

Whenever something is transferred from a working copy to
somewhere else via a patch, an added object must be captured
eventually anyway (no work to be avoided) and it must be done
early since the object could change while the save dialog is open.
When additions come from other versions instead, the artifacts
are immutable, so the reading can be safely postponed. Thus the
work of reading the artifact might be avoided if the addition is
excluded or the operation is cancelled.

Be more strict when comparing artifacts of different types.

Ref #130