Fix errors in merges
If working version had changes w.r.t. to the base and the incoming
version had no changes to the same artifact, a diff that reverted
the changes of the working version slipped into the merge.
Ensure that only SquotArtifactLoads are added in this way by not
considering the incoming artifact at all (it was not changed on the
incoming branch after all).

Merges of artifacts did not have the load status initialized at all
if the artifact was unloaded in the working copy.
The reason is that SquotArtifactMerge>>privateResolvedPatch
does not use the factory method of SquotArtifactModification.
In fact it is the only sender of #new to that class.

Ref #130