Fix #291: Pull created a merge commit even if a fast-forward was performed
The merge operation already determines whether a fast-forward
is possible. If yes, it updates the loadedVersion of the working copy.
If fast-forwarding is not possible, the merge operation will also add
the additional parent version, the pull doesn't need to do that.

Align pull more with SquitBrowser>>#mergeCommit:into:mergeTitle:saveTitle:saveMessageTemplate:.