Apply metadata differences in addition to object graph differences
...but not for all stores. Patching is currently a store-specific implementation.
There is no common code about artifact patching. This is probably because
patching the artifacts in a snapshot and then converting them
(materialize or serialize) would not allow to do precise updates to a store
when the differences are small (and mostly they are).

This is still only the "make it work" part of the solution, not the "make it right" part.

The manipulation of artifact metadata diffs or merges is not implemented yet.

Starting to shift responsiblilities of diff calculations away from the diff
and merge classes, which should only hold the differences. Merge classes
by design contain the diff calculation and they are hard to use in tests...

Start to rename mergeDiffWith:basedOn: to mergeWith:basedOn:, since diffs and
merges are two separate things.

Since artifact modifications can now modify the metadata without modifying
the object graph, add a no-changes guard to SquotObjectGraphDiff.

Curious: SquotLiveArtifact copyWithDiffApplied fails if the storeInfo is accessed
before the content is captured. Hence the order of setting these properties in
the copy were swapped.

outboundPointers always includes the class of the receiver, so an isEmpty check
makes no sense.

Ref #302