Be open for additional kinds of metadata members
Instead of hardcoding for Dictionary and variables of regular objects,
let mirrors and reflection determine which references belong together.

Change metadata with object references from IdentityDictionary
to Dictionary in the snapshot form. Distinct objects will be transformed
to different object names, so Dictionary is sufficient and will allow
comparing and merging of associations.

Ref #302