Populate artifact diffs and merges in the artifact itself
The compared objects know the data, so let them do the work.

Using an empty Object as the merge base in case there is no common
ancestor is only a workaround. nil can't be used currently because
the capturer will complain that primitive values should not get names assigned.

Extend object graph merging only so far that the tests are green again.

Ref #302