Be open to other kinds of reference diffs
SquotReferenceAdditionConflict is a modifed copy of SquotObjectAdditionConflict.
The reference classes are currently only used in metadata diffs and merges.
But they could be used in object diffs and merges. The object diff/addition/conflict
classes should be part of an object graph merge on the other hand.

Object graph merge
|- Object addition conflict
`- Object merge
    |- Object reference diff
    `- Object reference conflict

Since references are supposed to have only one target as of now,
there are no reference merges yet.

Ref #302