Merge back changes made by the repository store
Previously this was done in special cases of adding unloaded artifacts
and updating metadata of added artifacts.
If any conflicts arise with what was stored, stick to what is already loaded.
While this may in theory lead to "works on my machine" kinds of problems,
it is still better than silently corrupting some loaded objects during a save.

The metadata of an artifact put in the file store may actually change.
Previously this happened without notice because the artifact object
was silently modified (although it is treated as immutable otherwise)
by a lazy initialization of the serializerName property.
This also silently modified the working copy snapshot previously made,
masking the metadata change from the newly introduced merge.
Now, create a new copy of the artifact object to add the serializerName property.
Because of this, tests started to fail because now there are differences
between the added artifact and the stored artifact.
Only assert that the object graphs didn't change instead.

Because there is now a merge going on after saving,
tests with Text are broken due to the broken object patching.
See #305. Replaced Text by ValueHolder in tests and
marked some tests as expected to fail for now.

Note that object merges do not support value objects well yet.
Leave a notYetImplemented for that instead of a MessageNotUnderstood
for #name. Value objects will not get a name.

Ref #302