WIP: Use mirrors and walks in capturing and comparisons
Many former extension methods are now found in mirrors.
Mark the extension methods as deprecated.

Introduce mirrors for CompiledCode, Sets, classes, packages.

Reflect dictionary references as SquotCollectionElements that
refer to the Associations. SquotDictionaryEntry cannot capture
shared Associations between dictionaries. Moreover, it relies on
the keys of the original Associations, but these need to be
captured as well, so they will be different in the snapshot.

Exclude irrelevant instance variables again.

Since only the referrer objects know which role a referenced
object plays, the mirror of the referrer must be used to look up
the mirror for the referenced object. This is not yet implemented
everywhere.