WIP: Use mirrors and walks in materialization
Apply same principle as during capturing: deal with the references
of an object later after the object. For materialization this means:
create the empty instances first, then fill in the references whenever
their values are rematerialized, and after all references of an object
have been filled in, reactivate it.

Fix object shadow mirror: one #value message too much.
Arguably the selectors variablePartSlotWithIndex: and
instVarSlotNamed: are misleading because these methods do
not answer the slots, but their values.

Dictionaries are not correctly hashed after materialization yet.
The Associations are filled into the Dictionary before the keys are
filled in, so the hash values of the associations are not correct yet.
But the dictionary is not again reactivated/rehashed after the keys
have been filled into the Association objects.