WIP: Use mirrors and walks in materialization
Dictionary associations are now correctly hashed.
Complex Set elements are probably not though because
of the same problem: the hash changes while the elements
references are still being filled in, but at that time it has
already been added to the Set.

Had to resolve a chicken-and-the-egg problem for looking
up the mirror of an association: to do that it was previously
required to first add it to the dictionary, but that cannot work
because the key cannot be filled into the association yet
without the mirror. Solved it by allowing to supply the already
looked-up object to the mirror lookup method. The undecorating
of objects can also be done in better places because of that
change.

Patching is broken because of how the materializer has
been changed (especially materialize:in:).

At least prepare the patching test class for the new
materialization API usage (i. e. do not call rematerialize: to
materialize a whole subgraph).