WIP: Use mirrors and walks in materialization
Fix hashing in Sets. References can now indicate that the hash of their
value matters. References to Set elements and Dictionary associations
must do that, unless they stem from an IdentitySet or IdentityDictionary.

The reference to the key variable of a LookupKey is obviously also one
where the hash of the value (here: the key object) is relevant.

Cannot move this:
	remainingReferences removeKey: referrer.
...into the reactivate: method because objects whose references
are all filled in immediately are never entered into the
remainingReferences dictionary.

Patching is still broken (see previous commit message).