Improve keyboard efficiency when editing remotes
Select whole input texts by default.
Start with keyboard focus in the URL text box.
Pressing enter in the text box moves to the next box
or accepts the dialog, respectively.

There does not seem to be a way to deal with keyboard
focus in a UI toolkit agnostic manner, so this is Morphic-only.

Resolves #317