Use mirrors and references for patching
SquotNewObjectPatcher replaces SquotObjectPatcher.
A feature toggle is introduced, but actually the old patcher
does not work either. See previous commits about using
mirrors and walks in materialization.

Some tests are still failing. The reasons are that
- when the root of the object graph is replaced, the right
graph of the graph diff still contains the new root under the
name of the old root object,
- packages are not updated if they already exist and are
"materialized" instead of patched, although for the current
solution of treating packages, there should be no difference
between materializing and patching a package.

Add breadth first search mode to the graph walk because it
is faster when traversing graphs to find out the correct mirrors.
I do not understand why it is faster than depth first search
in this case, since at the time I measured this, the whole
graph would be traversed either way.

Allow to disable referrer tracking during walks.
It is not always needed.

It is faster to check during a walk whether an object 
has already been encountered before enqueuing it again
for the walk, rather than just ignoring the object if it is
enumerated again by the walk.

Add debugging facilities to SquotGraphWalk, such as a
history list to check which objects were traversed before.

#referenceDiffsWithDefaultOwners:reflectedIn:do: is like 

Classes and PackageInfos should not get an extra object name,
but if they are already present, they should be updated and
not replaced. Finally it is useful to have two different messages
for this: shouldHaveNameAssignedBySquot and squotShouldKeepIdentity.

This reverts commit 746de8a from 2018 (objectNamed:ifAbsent:),
which apparently just fixed symptoms, but not causes.

SquotImageToFileStoreTest>>testWithWorkspaceMorph
unexepectedly got fixed by the changes, even though the
documented cause of the failure was not fixed yet...

For objects decorated with a name, print the object itself first,
otherwise one has to scroll too far right to see the actual value.

Break when attempting to reflect on a SquotNameDecorator
instead of the decorated object directly. Usually this indicates
a missing send of #undecorated.