Fix: new object in SquotObjectReplacement was not materialized
Patched side of GraphDiff had wrong startName, thus the old object
was found. So it was incorrectly deduced that the object already exists.

SquotObjectReplacement is the single special case where it looks
like an object changes the name, or to be correct: where the original
side of a difference has a different name than the patched side of it.
This requires some special cases elsewhere as well, unfortunately.

Ideally, there would be no need for SquotObjectReplacement. It would
require that object graph capturing starts from a reference, not from
a start object. But this requires even further refactoring. Maybe later...

Moved name assigning from the graph to the diff. If the name must
change, the diff will know.

The workaround with SquotObjectGraphMirror, SquotObjectGraphKickoff etc.
requires that the objects are registered sooner in the target graph.
Therefore do the updating of references to treated objects after assigning
the name to the treated object.

Use TSquotDiff Trait method instead of SquotObjectDiff>>applyToShadowNamed:inGraph: