Fix materialization if new object existed in a different state already
Rework materialization procedure. First capture the current graph
again to get updated shadows, then compute a graph diff to the
desired state of the graph. Ensure that the target graph knows
all the relevant objects that already exist. Then apply the diff.

If the start object is replaced and the new object did not
exist in the destination graph yet, all its references must
also be materialized. Have a SquotObjectAddition in the graph
diff in addition to the SquotObjectReplacement because the
diff of the value reference could not be looked up under the
name of the new object, since the replacement is filed under
the name of the old object.

Prevent messages not understood by nil due to object graphs
not having a live start object yet.

Do not attempt to look up references in objects of a different kind
for a replacement.

Package materialization does still not work. A SquotObjectAddition
does not apply package changes like a SquotPackageDiff would.